pyproject.toml
[tool.poetry]
name = "gradio-fastapi-auth0"
version = "0.1.0"
description = ""
authors = ["ren-surf <ren.nakamura.engineer@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.5"
itsdangerous = "^2.2.0"
authlib = "^1.3.2"
httpx = "^0.28.0"
gradio = "^5.7.1"
python-dotenv = "^1.0.1"
ruff = "^0.8.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "F",   # pyflakes
    "B",   # flake8-bugbear
    "I",   # isort
]